fabric: move blitz hop allocator into topology_mapper_utils
Relocate assign_non_colliding_hops() out of blitz_decode_pipeline into the shared
experimental fabric utils (topology_mapper_utils.{hpp,cpp}, namespace
tt::tt_metal::experimental::tt_fabric). It is a generic 'distinct representatives'
backtracking solver, so it belongs in fabric utils rather than the blitz pipeline's
public surface: the blitz internal header is unchanged, build_pipeline_allocation_from_topology
calls the util, and the CPU-only BlitzDecodePipelineAssignment unit tests include the
utils header directly (no forward-declaration). All 7 unit tests pass.
