+++ tests/tt_metal/tt_fabric/fabric_router/test_routing_tables.cpp
@@ -12,7 +12,6 @@
 #include <unordered_set>
 #include <utility>
 #include <yaml-cpp/yaml.h>
-#include <tt-metalium/experimental/internal/blitz_decode_pipeline.hpp>
 
 #include "fabric_fixture.hpp"
 #include "t3k_mesh_descriptor_chip_mappings.hpp"
@@ -29,6 +28,7 @@
 #include <tt-metalium/distributed_context.hpp>
 #include <tt-logger/tt-logger.hpp>
 #include <fmt/format.h>
+#include <tt-metalium/experimental/fabric/topology_mapper_utils.hpp>
 
 namespace {
 
@@ -2259,7 +2259,7 @@ namespace blitz_assign_tests {
 
 using ::tt::tt_fabric::FabricNodeId;
 using ::tt::tt_fabric::MeshId;
-using ::tt::tt_metal::internal::blitz::detail::assign_non_colliding_hops;
+using ::tt::tt_metal::experimental::tt_fabric::assign_non_colliding_hops;
 using HopPair = std::pair<FabricNodeId, FabricNodeId>;
 
 FabricNodeId node(std::uint32_t mesh, std::uint32_t chip) { return FabricNodeId(MeshId{mesh}, chip); }
